v2.2.2 change vacations days from int to double

+++ src/main/java/com/countmyh/model/MonthNote.java
@@ -1,3 +1,3 @@
 package com.countmyh.model;
 
-public record MonthNote(int year, int month, int holidays, String observation) {}
+public record MonthNote(int year, int month, double holidays, String observation) {}
